requirements.txt
matplotlib
nltk
numpy
sklearn
tensorboard==1.7.0
tensorflow-gpu==1.7.0
tqdm
